requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
